refactor: support 관련 타입 구조 개선 (#115)
