+++ src/components/support/chatRoom/ChatHeader.tsx
@@ -1,7 +1,13 @@
-import type { ChatHeaderProps } from '@/types/support';
+import type { ChatHeaderInfoType } from '@/types/support';
 import { BsTelephoneXFill } from 'react-icons/bs';
 import { useChatStatusStore } from '@/store/chatStatusStore';
 import Badge from '@/components/common/Badge';
+
+interface ChatHeaderProps {
+  chatInfo: ChatHeaderInfoType;
+  onSupportEnd: (sessionId: number) => void;
+  forcedRefresh?: boolean;
+}
 
 export default function ChatHeader({
   chatInfo,
